Split off measurement recording from active and noop executions.
