+++ src/main/java/com/iopipe/IOpipeExecution.java
@@ -32,21 +32,6 @@ public abstract class IOpipeExecution
 	protected final boolean coldstart;
 	
 	/**
-	 * Performance entries which have been added to the measurement, this
-	 * field is locked since multiple threads may be adding entries.
-	 */
-	private final Set<PerformanceEntry> _perfentries =
-		new LinkedHashSet<>();
-	
-	/** Custom metrics that have been added, locked for thread safety. */
-	private final Set<CustomMetric> _custmetrics =
-		new LinkedHashSet<>();
-	
-	/** Labels which have been added, locked for threading. */
-	private final Set<String> _labels =
-		new LinkedHashSet<>();
-	
-	/**
 	 * Initializes the base execution.
 	 *
 	 * @param __cold Is this a cold start?
@@ -56,6 +41,16 @@ public abstract class IOpipeExecution
 	{
 		this.coldstart = __cold;
 	}
+	
+	/**
+	 * Adds a single performance entry to the report.
+	 *
+	 * @param __e The entry to add to the report.
+	 * @throws NullPointerException On null arguments.
+	 * @since 2018/01/19
+	 */
+	public abstract void addPerformanceEntry(PerformanceEntry __e)
+		throws NullPointerException;
 	
 	/**
 	 * Returns the configuration used to initialize the service.
@@ -76,6 +71,69 @@ public abstract class IOpipeExecution
 	public abstract Context context();
 	
 	/**
+	 * Adds a single custom metric to the report.
+	 *
+	 * @param __cm The custom metric to add.
+	 * @throws NullPointerException On null arguments.
+	 * @since 2018/01/20
+	 */
+	public abstract void customMetric(CustomMetric __cm)
+		throws NullPointerException;
+	
+	/**
+	 * Adds the specified custom metric with a string value.
+	 *
+	 * Custom metric names are limited to the length specified in
+	 * {@link IOpipeConstants#NAME_CODEPOINT_LIMIT}.
+	 *
+	 * @param __name The metric name.
+	 * @param __sv The string value.
+	 * @throws NullPointerException On null arguments.
+	 * @since 2018/01/30
+	 */
+	public abstract void customMetric(String __name, String __sv)
+		throws NullPointerException;
+	
+	/**
+	 * Adds the specified custom metric with a long value.
+	 *
+	 * Custom metric names are limited to the length specified in
+	 * {@link IOpipeConstants#NAME_CODEPOINT_LIMIT}.
+	 *
+	 * @param __name The metric name.
+	 * @param __lv The long value.
+	 * @throws NullPointerException On null arguments.
+	 * @since 2018/01/30
+	 */
+	public abstract void customMetric(String __name, long __lv)
+		throws NullPointerException;
+	
+	/**
+	 * Returns a copy of the custom metrics which were measured.
+	 *
+	 * @return The custom metrics which were measured.
+	 * @since 2018/03/15
+	 */
+	public abstract CustomMetric[] getCustomMetrics();
+	
+	/**
+	 * Returns all of the labels which have been declared during the
+	 * execution.
+	 *
+	 * @return The labels which have been declared during execution.
+	 * @since 2018/04/11
+	 */
+	public abstract String[] getLabels();
+	
+	/**
+	 * Returns a copy of the performance entries which were measured.
+	 *
+	 * @return The performance entries which were measured.
+	 * @since 2018/03/15
+	 */
+	public abstract PerformanceEntry[] getPerformanceEntries();
+	
+	/**
 	 * Returns the object which was used as input for the method being
 	 * executed, {@code null} will be returned if it was not passed or is not
 	 * known.
@@ -85,6 +143,19 @@ public abstract class IOpipeExecution
 	 * @since 2018/04/16
 	 */
 	public abstract Object input();
+	
+	/**
+	 * Adds a single label which will be passed in the report.
+	 *
+	 * Labels are limited to the length specified in
+	 * {@link IOpipeConstants#NAME_CODEPOINT_LIMIT}.
+	 *
+	 * @param __s The label to add.
+	 * @throws NullPointerException On null arguments.
+	 * @since 2018/04/11
+	 */
+	public abstract void label(String __s)
+		throws NullPointerException;
 	
 	/**
 	 * This returns an instance of a plugin based on the class type of its
@@ -130,95 +201,6 @@ public abstract class IOpipeExecution
 	public abstract long startTimestamp();
 	
 	/**
-	 * Adds a single performance entry to the report.
-	 *
-	 * @param __e The entry to add to the report.
-	 * @throws NullPointerException On null arguments.
-	 * @since 2018/01/19
-	 */
-	public final void addPerformanceEntry(PerformanceEntry __e)
-		throws NullPointerException
-	{
-		if (__e == null)
-			throw new NullPointerException();
-		
-		// Multiple threads could be adding entries
-		Set<PerformanceEntry> perfentries = this._perfentries;
-		synchronized (perfentries)
-		{
-			// Performance entry was defined, so just say that the plugin was
-			// used for tracing data
-			this.label("@iopipe/plugin-trace");
-			
-			perfentries.add(__e);
-		}
-	}
-	
-	/**
-	 * Adds a single custom metric to the report.
-	 *
-	 * @param __cm The custom metric to add.
-	 * @throws NullPointerException On null arguments.
-	 * @since 2018/01/20
-	 */
-	public final void customMetric(CustomMetric __cm)
-		throws NullPointerException
-	{
-		if (__cm == null)
-			throw new NullPointerException();
-		
-		// Multiple threads can add metrics at one time
-		Set<CustomMetric> custmetrics = this._custmetrics;
-		synchronized (custmetrics)
-		{
-			if (!__cm.name().startsWith("@iopipe/"))
-				this.label("@iopipe/metrics");
-			
-			custmetrics.add(__cm);
-		}
-	}
-	
-	/**
-	 * Adds the specified custom metric with a string value.
-	 *
-	 * Custom metric names are limited to the length specified in
-	 * {@link IOpipeConstants#NAME_CODEPOINT_LIMIT}.
-	 *
-	 * @param __name The metric name.
-	 * @param __sv The string value.
-	 * @throws NullPointerException On null arguments.
-	 * @since 2018/01/30
-	 */
-	public final void customMetric(String __name, String __sv)
-		throws NullPointerException
-	{
-		if (__name == null || __sv == null)
-			throw new NullPointerException();
-		
-		this.customMetric(new CustomMetric(__name, __sv));
-	}
-	
-	/**
-	 * Adds the specified custom metric with a long value.
-	 *
-	 * Custom metric names are limited to the length specified in
-	 * {@link IOpipeConstants#NAME_CODEPOINT_LIMIT}.
-	 *
-	 * @param __name The metric name.
-	 * @param __lv The long value.
-	 * @throws NullPointerException On null arguments.
-	 * @since 2018/01/30
-	 */
-	public final void customMetric(String __name, long __lv)
-		throws NullPointerException
-	{
-		if (__name == null)
-			throw new NullPointerException();
-		
-		this.customMetric(new CustomMetric(__name, __lv));
-	}
-	
-	/**
 	 * Adds multiple custom metrics in a single bulk operation.
 	 *
 	 * Parameters which are {@code null} are ignored.
@@ -231,20 +213,11 @@ public abstract class IOpipeExecution
 		// Do nothing
 		if (__cms == null)
 			return;
-		
-		// Bulk add all the custom metrics under a single lock
-		Set<CustomMetric> custmetrics = this._custmetrics;
-		synchronized (custmetrics)
-		{
-			for (CustomMetric cm : __cms)
-				if (cm != null)
-				{
-					if (!cm.name().startsWith("@iopipe/"))
-						this.label("@iopipe/metrics");
-					
-					custmetrics.add(cm);
-				}
-		}
+			
+		// Add all metrics
+		for (CustomMetric cm : __cms)
+			if (cm != null)
+				this.customMetric(cm);
 	}
 	
 	/**
@@ -276,78 +249,6 @@ public abstract class IOpipeExecution
 	public final boolean isColdStarted()
 	{
 		return this.coldstart;
-	}
-	
-	/**
-	 * Returns a copy of the custom metrics which were measured.
-	 *
-	 * @return The custom metrics which were measured.
-	 * @since 2018/03/15
-	 */
-	public final CustomMetric[] getCustomMetrics()
-	{
-		Collection<CustomMetric> custmetrics = this._custmetrics;
-		synchronized (custmetrics)
-		{
-			return custmetrics.<CustomMetric>toArray(
-				new CustomMetric[custmetrics.size()]);
-		}
-	}
-	
-	/**
-	 * Returns all of the labels which have been declared during the
-	 * execution.
-	 *
-	 * @return The labels which have been declared during execution.
-	 * @since 2018/04/11
-	 */
-	public final String[] getLabels()
-	{
-		Set<String> labels = this._labels;
-		synchronized (labels)
-		{
-			return labels.<String>toArray(new String[labels.size()]);
-		}
-	}
-	
-	/**
-	 * Returns a copy of the performance entries which were measured.
-	 *
-	 * @return The performance entries which were measured.
-	 * @since 2018/03/15
-	 */
-	public final PerformanceEntry[] getPerformanceEntries()
-	{
-		Collection<PerformanceEntry> perfentries = this._perfentries;
-		synchronized (perfentries)
-		{
-			return perfentries.<PerformanceEntry>toArray(
-				new PerformanceEntry[perfentries.size()]);
-		}
-	}
-	
-	/*
-	 * Adds a single label which will be passed in the report.
-	 *
-	 * Labels are limited to the length specified in
-	 * {@link IOpipeConstants#NAME_CODEPOINT_LIMIT}.
-	 *
-	 * @param __s The label to add.
-	 * @throws NullPointerException On null arguments.
-	 * @since 2018/04/11
-	 */
-	public final void label(String __s)
-		throws NullPointerException
-	{
-		if (__s == null)
-			throw new NullPointerException();
-		
-		// Add it
-		Set<String> labels = this._labels;
-		synchronized (labels)
-		{
-			labels.add(__s);
-		}
 	}
 	
 	/**
